Add a helper to manually invalidate contact cache
In our game, sometimes an object can be forced up against a wall by a force and then a hole is cut into the wall it is on, and it doesn't update as it still thinks its it's making contact, even when forcing an Activate, etc.

We know this ahead of time so we can manually invalidate the contact cache for the object that was touching.