Add TrackContactsForThisBodyPair to ValidateResult
Adds a new validate result, in which the object's collision can be tracked for contact, without actually performing any constraints, etc on the object.

This essentially makes it work like a sensor, following the validation step, which can be useful if the body only wants to act as a sensor to some objects and not others.

It's also useful if your game/engine expects to get Start/EndTouch callbacks even when collision is disabled from an object from testing if two narrow-phase bodies should collide.