Improved handling of hulls with very acute angles (#179)
When points were nearly co-planar (but not co-planar enough for the algorithm to detect a 2D hull) it was possible that a point that was within tolerance to the plane but outside of the polygon was discarded. Now we calculate the distance to the edges and add the point to the hull at the end.