Discarding slivers when building a mesh shape (#606)
When slivers are added to a mesh shape, collision detection with these triangles will result in triangle_normal being zero in CollideConvexVsTriangles::Collide. ActiveEdges::FixNormal will sometimes find that the hit is an interior hit and return the triangle normal as penetration axis. This penetration axis which will be normalized in PhysicsSystem::ReductionCollideShapeCollector and result in a NaN causing the simulation to hang indefinitely the next time step. To fix this, we remove slivers when sanitizing a MeshShapeSettings.

This should fix godot-jolt/godot-jolt#485.