Respecting the mIsSensor flag for CCD bodies (#637)
When a body with the LinearCast motion quality type is converted to a sensor in the OnContactAdded callback, the body would still trigger a collision response and bounce off. Now the mIsSensor flag is respected and collision response handling is skipped. Note that since we only find the first hit for the LinearCast motion type, this makes it possible to penetrate another body that is right behind the body that got converted to a sensor.