Added constraint priority to control the order of evaluation of constraints (#638)
This functionality can be used to make some constraints more important than others (because they have higher priority and are evaluated later) but it can also be used when creating a deterministic simulation between client and server where the client doesn't simulate all objects so is missing some of the constraints. By setting a unique constraint priority on every constraint it will still be possible to create a deterministic simulation for a subset of the constraints. See discussion at #611.