Bugfix: Character reports to be on ground for one more frame after jumping against a wall
The wall will report a sliding contact and if we're not supported but have a sliding contact, we will attempt to move downwards to see if we're not actually supported by multiple steep slopes that form a bowl. We should skip this check if we're actually moving away from the ground as we'll not be supported then.