Fixed issue in contact normal calculation in EPA algorithm (#683)
If, during the building of the hull, the hull becomes invalid (e.g. triangles facing the origin) there is a chance that the last processed triangle is actually not close to the origin at all (the support function returns a high distance). In order to reduce the chance of using this triangle as the contact point and normal we now compare the before last triangle with the last triangle to try to detect this case.

Fixes #671