Fixing flipped normals en EPA penetration depth algorithm (#691)
With a hull consisting of two triangles that are back to back, it is possible that the side on which the origin is is misclassified for one of the two triangles (due to floating point precision issues). The resulting penetration axis (= triangle normal) will be flipped because of that resulting in a penetration axis that pushes the objects together rather than separating them. In order to avoid this, we do an additional check when we detect a failure in building a correct hull to see if the penetration is less in the flipped normal direction and if it is we flip the penetration axis.

Fixes #671