Fixed race condition in narrow phase query
If a body is removed between the broad phase detecting an overlap and the narrow phase locking the body, callbacks may be called on a body that has already been removed (it will still be a valid body because otherwise the lock would have failed). We ensure that the callbacks can assume that the body is still in the broadphase. The collision collector cannot make this assumption (as before).