Made GetProcessorTicksPerSecond more reliable and removed some dependencies on it (#706)
GetProcessorTicksPerSecond was using /proc/cpuinfo on Linux. The returned value is not accurate on ARM and apparently also not on new x86 AMD CPU's. Using std::chrono now to measure the number of ticks per second and updating it every frame to ensure that it stays reasonably accurate if the ticks/second value changes with e.g. CPU power states. Note that this function is only used for profiling information, so even if the value is wrong, the simulation will not be affected.

See: https://x.com/phys_ballsocket/status/1701933316654383164?s=20