Fixed character jittering issue when moving down on elevator (#708)
Because the character update and physics tick happen at different moments, we use the current velocities of the bodies to predict the state at the next physics tick and move the character accordingly. For a character on an elevator moving down this meant that it would predict a position that's correct for the next physics tick, but for the current state is inside the elevator a bit. The collision check that does a final check after the movement has been calculated (GetFirstContactForSweep) uses the current position of the physics bodies and it will prevent the character from moving into the elevator.

There was an issue in the way GetFirstContactForSweep handled character padding which meant it didn't actually detect a collision with the elevator (the sweep was not long enough) so the character would be allowed to sink in the elevator. Every X frames the character would have sunken into the elevator enough for GetFirstContactForSweep to detect a collision without the character padding and it would suddenly shift the character up with this padding. This resulted in a visual jitter.

Note that there are still some artifacts that are really hard to avoid:

* When moving down, the player will visually hover a little bit above the elevator because the collision check doesn't take into account that the elevator will move next tick (we have to be conservative because we cannot guarantee that it will actually move with the velocity we see in the character update, the velocity of a dynamic object cannot be predicted without running the full physics tick)
* When transitioning from moving down to up, the player will briefly lose contact with the floor because it sees the updated elevator velocity and moves clear of the elevator resulting in the InAir state.

Should fix #703