Changed AlignedAllocate to default to posix_memalign (#738)
I'm in the process of implementing iOS support for Godot Jolt, and it turns out that aligned_alloc is only available on iOS >=13, as seen here. According to that header it also appears that aligned_alloc is not available on macOS versions prior to 10.15 (Catalina). What is available however is posix_memalign.

I figured instead of adding even more platform ifdef's to AlignedAllocate I would just simplify things a bit and always default to posix_memalign on anything but Windows, as this seems to be available on Android as well. Although, I'm not sure how this change fares on that one blue platform.

Also note that posix_memalign technically returns an error code, but seeing as how none of the other allocation functions had any error-handling I opted to just ignore it.