Adds a sample project to test the virtual character determinism.
This commit introduce a basic rewinding mechanism, that allows to record
a series of input and replay them. This mechanism can be used to test the
VirtualCharacter determinism on many situations.

When playing, is now possible to press `Y` to start recording the inputs
(press `Y` again to stop it), and then by pressing `U` it's possible to replay
those inputs and validate the VirtualCharacter state on each frame.
On the UI, will be displayed the list of frames which the state differs from the
one originally recorded so the rewinding was not deterministic.