Adds a flag to enable assertion in non debug builds.
Unreal Engine has two type of compilation targets DebugGame and Development. When compiling in DebugGame, despite the name, the produced binary is not a target Debug but it's still target Release but without optimizations.
This proposed change allows to enable the assertion even for non debug builds, so we can still use assertions even in Unreal Engine.