Replace std::vector with custom implementation (#1089)
std::vector memsets the entire data block to 0 before doing anything else and there doesn't appear to be a reasonable way around it. This was actually costing quite a bit of time when deserializing big meshes/height fields. This change replaces std::vector with a custom Array class, but it can still be turned off by enabling the JPH_USE_STD_VECTOR define (or the USE_STD_VECTOR cmake option).

This change also adds a Reallocate function that needs to be implemented when you override the memory allocators and a reallocate function that you need to implement if you have a custom array allocator. The behavior is the same as the C realloc function. It is used to reallocate a block of memory for simple types instead of always going through a alloc, copy, free cycle.