FixedSizeFreeList: Reorder variable layout to reduce false sharing/syncs
Currently, the bookkeeping variables used in `Get()` calls reside in
the same cache line as the atomic variables which are modified upon
allocation. If one thread allocates a new object and another thread
calls `Get()`, a cross-thread cache sync has to occur, causing a stall
and reducing performance.

By reordering variables in the allocator, we can move the variables used
in `Get()` to a different cache line, so threads don't need to sync
for object reads after an allocation occurs.