QuadTree / FixedSizeFreeList: Reorder variable layout to reduce false sharing & thread syncs (#1136)
* FixedSizeFreeList: Reorder variable layout to reduce false sharing/syncs

Currently, the bookkeeping variables used in `Get()` calls reside in
the same cache line as the atomic variables which are modified upon
allocation. If one thread allocates a new object and another thread
calls `Get()`, a cross-thread cache sync has to occur, causing a stall
and reducing performance.

By reordering variables in the allocator, we can move the variables used
in `Get()` to a different cache line, so threads don't need to sync
for object reads after an allocation occurs.

* QuadTree: Reorder variable layout to reduce false sharing/thread syncs

Testing across thread counts on modern desktop x86 processors using the built-in PerformanceTest (compiled with LLVM 17 via Clang-CL on Windows), this seems to give around a 5% average improvement in the general case. On other workloads I've tested which are heavily memory-bound, I've seen improvements of up to 15% at specific thread counts. This has also been tested on Apple silicon with no regressions found.