Improved enhanced active edge detection (#1162)
Turning off the cached active edge flags when enhanced active edge detection is on. A collision with an inactive edge will result in its normal being aligned with the face. This means it is not properly filtered out by the enhanced active edge detection algorithm and can cause an incorrect simulation.

See #1155