Added Shape::MakeScaleValid function (#1172)
This function will take a scale vector and check it against the scaling rules for the shape. If it is not valid, it will return a scale that is close to the provided scale which is valid. For e.g. sphere shapes it means that the output is always a uniform scale, for a cylinder X and Z are averaged.
Bugfix: The IsValidScale will now also return false when one of the scale components is zero.