Fixed crash when soft body collides with height field
When a height field is created where SampleCount / BlockSize is not a power of 2 and the soft body touched the right or bottom border of the height field, the application would crash. The VisitRangeBlock implementation for CollideSoftBodyVertices did not ignore inside out bounding boxes and tried to read non-existing height field samples.

See report: https://github.com/godot-jolt/godot-jolt/issues/951#issuecomment-2321166739