SoftBodyManifold now returns sensor contacts separately (#1276)
Before this change, there was a limit of a single colliding body per soft body vertex. If the closest body happened to be a sensor this effectively disabled the collision with the world and caused artifacts. We can now also detect multiple sensor contacts per soft body and they are returned through a new interface SoftBodyManifold::GetSensorContactBodyID.

* Shape::CollideSoftBodyVertices is no longer using SoftBodyVertex but CollideSoftBodyVertexIterator which can write to other things than a SoftBodyVertex
* Removed delta time and displacement due to gravity from Shape::CollideSoftBodyVertices
* Added soft body vs sensor test that doesn't have a limit of 1 sensor per soft body

Fixes #1016