Check Jolt with TSAN in CI
This commit adds a new GitHub Actions workflow that checks
Jolt with ThreadSanitizer under Ubuntu & Clang. The workflow
is triggered on every push to the repository.

This commit also replaces usage of atomic_thread_fences in
Reference.h and JobSystem.h with regular atomic ops when
building under TSAN, as it does not support fences and unlikely
ever will do so.