Fixed TSAN false positive
TSAN detects a race between BodyManager::AddBodyToActiveBodies coming from PhysicsSystem::ProcessBodyPair and Body::GetIndexInActiveBodiesInternal, however we process the active bodies in batches using an atomic BodyManager::mNumActiveBodies with memory order acquire, so we know that the body is in a consistent state by the time we read the active index. Workaround is to use an atomic for mIndexInActiveBodies that we always read/write in relaxed mode.